Fix/157619922/limit move integration tests folder
[Starts #157619922]
